UNIC-2005: replace double with decimal

+++ dags/sql/indicateurs_sip/tables/toutes_les_infections_schema.sql
@@ -30,6 +30,6 @@ CREATE TABLE IF NOT EXISTS indicateurs_sip.toutes_les_infections (
    INDICATEUR_CIBLE Int,
    Calcul_ValeurCoefficient Int,
    Calcul_Denominateur Int,
-   Calcul_Cible Double,
+   Calcul_Cible DECIMAL,
    LISTDENOMINATEUR_ANNEE VARCHAR
 );
